+++ frontend/package.json
@@ -23,6 +23,7 @@
   },
   "dependencies": {
     "@aadeshk/medium-common": "^1.0.2",
+    "@types/quill": "^2.0.14",
     "axios": "^1.6.8",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
